% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARM.IWW-package.R
\docType{package}
\name{ARM.IWW-package}
\alias{ARM.IWW}
\alias{ARM.IWW-package}
\title{ARM.IWW: Multi-site stochastic rainfall model based on an alternating renewal process}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ross Pidoto \email{pidoto@iww.uni-hannover.de} (\href{https://orcid.org/0000-0001-7471-2615}{ORCID})

}
\keyword{internal}
