% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_copulas.R
\name{fit_copulas}
\alias{fit_copulas}
\title{Title}
\usage{
fit_copulas(
  StationID,
  AddNoise = TRUE,
  DateFrom = NULL,
  DateTo = NULL,
  ThresholdNAs = 0.1,
  saveObjects = TRUE,
  ModelType = get_model_type()
)
}
\arguments{
\item{ModelType}{}
}
\description{
Title
}
