% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_marginals.R
\name{fit_marginals}
\alias{fit_marginals}
\title{Fit marginal probabilty distributions}
\usage{
fit_marginals(
  StationID,
  AddNoise = TRUE,
  DateFrom = NULL,
  DateTo = NULL,
  ThresholdNAs = 0.1,
  saveObjects = TRUE,
  ModelType = get_model_type()
)
}
\arguments{
\item{StationID}{a character string of the station to process.}

\item{AddNoise}{logical, should noise be added to WSD and DSD?}

\item{DateFrom}{restrict events >= DateFrom. Optional.}

\item{DateTo}{restrict events <= DateTo. Optional.}

\item{ThresholdNAs}{numeric, above which proportion of NA values within an event should the event be excluded?}

\item{saveObjects}{logical, should marginals be saved to disk?}
}
\value{
A list containing the marginal distribution parameters for each event variable for summer and winter.
}
\description{
Fits marginal probability distributions to rainfall event variables for summer and winter separately.
}
