% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_external.R
\name{gen_external}
\alias{gen_external}
\title{Title}
\usage{
gen_external(
  StationID,
  Years,
  StartYear = 1900,
  Realisation = NA,
  RegEmpCopula = StationID,
  saveObjects = TRUE,
  ModelType = get_model_type()
)
}
\arguments{
\item{ModelType}{}
}
\description{
Title
}
