% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_internal.R
\name{gen_internal}
\alias{gen_internal}
\title{Title}
\usage{
gen_internal(
  StationID,
  Years,
  Realisation = NA,
  Alpha = 1/3,
  saveSpatialObjects = TRUE,
  saveObjects = TRUE,
  ModelType = get_model_type()
)
}
\arguments{
\item{ModelType}{}
}
\description{
Title
}
