% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Title}
\usage{
get_events(
  PcpTS,
  StationID,
  SmallEventsMode = FALSE,
  WSImin = 0.0099,
  DSDMax = 6 * 7 * 24 * 60,
  saveRObjects = TRUE,
  saveText = FALSE,
  includeEventTS = FALSE
)
}
\arguments{
\item{includeEventTS}{}
}
\description{
Title
}
