% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_small_events.R
\name{insert_small_events}
\alias{insert_small_events}
\title{Title}
\usage{
insert_small_events(
  StationID,
  ObsDateFrom = NULL,
  ObsDateTo = NULL,
  Years,
  Realisation = NA,
  saveObjects = TRUE,
  ModelType = get_model_type()
)
}
\arguments{
\item{ModelType}{}
}
\description{
Title
}
